+++ core/include/js_block.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <stdbool.h>
+#include "js_types.h"
+
+jsval_t js_block(JS *js, bool create_scope);
