+++ core/include/js_err.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "js_types.h"
+
+jsval_t js_err(JS *js, const char *fmt, ...);
