+++ core/include/js_include.h
@@ -1,0 +1,21 @@
+#pragma once
+#include "js_basic.h"
+#include "js_block.h"
+#include "js_entity.h"
+#include "js_err.h"
+#include "js_ffi.h"
+#include "js_func.h"
+#include "js_gc.h"
+#include "js_logic.h"
+#include "js_mem.h"
+#include "js_obj.h"
+#include "js_op.h"
+#include "js_parser.h"
+#include "js_prop.h"
+#include "js_scope.h"
+#include "js_stmt.h"
+#include "js_string.h"
+#include "js_stringnify.h"
+#include "js_types.h"
+#include "js_value_size.h"
+#include "js_variable.h"
