+++ core/include/js_stmt.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "js_types.h"
+
+jsval_t js_block_or_stmt(JS *js);
+jsval_t js_stmt(JS *js, uint8_t etok);
