+++ core/include/js_variable.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "js_types.h"
+
+jsval_t js_let(JS *js);
