+++ debug/js_debug.h
@@ -1,14 +1,15 @@
 #pragma once
 
 
-#define JS_DEBUG 1u
+#ifdef JS_CFG_DEBUG
 
-#ifdef JS_DEBUG
 #define JS_DUMP js_dump
 #define JS_DEBUGF printf
+void js_dump(struct js *js);
+
 #else
+
 #define JS_DUMP(...)   /**/
 #define JS_DEBUGF(...) /**/
-#endif
 
-void js_dump(struct js *js);
+#endif
